requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anaconda-anon-usage
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
archspec
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
boltons
Brotli
certifi
cffi
charset-normalizer
comm==0.2.2
conda
conda-content-trust
conda-libmamba-solver
conda-package-handling
conda_package_streaming
contourpy==1.2.1
cryptography
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dgl==2.4.0+cu121
distro
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.14.0
flexcache==0.3
flexparser==0.4
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2024.5.0
grpcio==1.64.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.5
httpx==0.27.0
hydra-core==1.3.2
idna
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.5.0
json5==0.9.25
jsonpatch
jsonpointer==2.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab-language-pack-zh-CN==4.2.post1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
libmambapy
lightning==2.5.1.post0
lightning-utilities==0.14.3
loguru==0.7.3
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
menuinst
mistune==3.0.2
-e git+https://github.com/choderalab/modelforge.git@d64c58088bc86ecea6b5adf195c9d65f138ba962#egg=modelforge
mpmath==1.3.0
multidict==6.4.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openff-toolkit
openff-units
-e git+https://github.com/openforcefield/openff-utilities.git@591a3f39af71ca3371972be3356628a15b6633ec#egg=openff_utilities
overrides==7.7.0
packaging
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
Pint==0.24.4
platformdirs
pluggy
prometheus_client==0.20.0
prompt_toolkit==3.0.45
propcache==0.3.1
protobuf==5.27.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycosat
pycparser
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.18.0
pyparsing==3.1.2
PySocks
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
ruamel.yaml
scikit-learn==1.6.1
scipy==1.15.3
Send2Trash==1.8.3
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
supervisor==4.2.5
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.3.0
torch
torchdata==0.9.0
torchmetrics==1.7.1
torchvision
tornado==6.4
tqdm
traitlets==5.14.3
truststore
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
urllib3
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.11
yarl==1.20.0
zstandard